refactored the pipeline builder
