--- ezgfx/src/bind/uniform.rs
@@ -1,19 +1,0 @@
-use super::*;
-
-pub struct Uniform
-{
-
-}
-
-impl Bind for Uniform
-{
-    fn ty() -> wgpu::BindingType
-    {
-        wgpu::BindingType::UniformBuffer { dynamic: false }
-    }
-
-    fn create(&self, layout: wgpu::BindGroupLayout) -> wgpu::BindingResource
-    {
-        todo!()
-    }
-}
