+++ ezgfx/src/lib.rs
@@ -1,8 +1,7 @@
-// mod pipeline;
-mod builder;
-//mod bind;
+mod pipeline;
+mod bind;
 mod ctx;
 
-// pub use pipeline::*;
-// pub use bind::*;
+pub use pipeline::*;
+pub use bind::*;
 pub use ctx::*;
